requirements.txt
numpy~=1.21.2
torch~=1.9.0
opencv-python~=4.5.3.56
Pillow~=8.3.2
torchvision~=0.10.0
pandas~=1.3.3
tqdm~=4.62.2
SimpleITK~=2.1.1
scikit-learn~=0.24.2
PyYAML~=5.4.1
pydicom~=2.2.1
omegaconf~=2.1.1
pytorch-lightning~=1.4.7
segmentation_models_pytorch~=0.2.0